Add support for backup "standby" hosts
Suppose you want to temporarily shut down an app running on server A
and move it to B. This requires running `stop` on A and `start` on B,
but without deploying this will start an old version on machine B. This
is easy to fix by altering the sbt settings to deploy on B instead and
running `deploy`. But not everybody can run `deploy` - system
specialists might only know about initctl and that's it.

To this end, this creates a `backupHosts` directive that will cause
libraries to be rsynced to those host **but not** started. So some
sysadmin can readily go and shut down the app on `hosts` and then
start them on `backupHosts`, and they will start the same version.